LayoutTests/media/fullscreen-content-offset.html

<html>
<head>
    <meta name="fullscreen-content-offset" content="width=device-width, initial-scale=1">
    <title>Test that element is visible on screen when scroll events are fired while entering fullscreen</title>
    <style>        
        .video-grandparent {
            display: inline-block;
        }
        .video-container {
            background-color: yellow;
            color: white;
            width: 400px;
            height: 400px;
            margin: 1em;
        }
    </style>

    <script src="video-test.js"></script>
    <script>
        let observer;
        let fullscreenElement;
    
        function logElementFailure(element, msg)
        {
            const bounds = element.getBoundingClientRect();
            const location = `[${bounds.left}, ${bounds.top}, ${bounds.right}, ${bounds.bottom}]`;
            const viewportLocation = `[${window.visualViewport.offsetLeft}, ${window.visualViewport.offsetTop}]`;
            const size = `${bounds.right - bounds.left}x${bounds.bottom - bounds.top}`;

            consoleWrite(`<span style='color:red'>FAIL: ${msg} when element is occluded, viewport scroll = ${viewportLocation}, element location = ${location}, element size = ${size}</span>`);
        }
    
        function calculateVisibilityPercentage(bounds)
        {
            const boundsArea = (bounds.right - bounds.left) * (bounds.bottom - bounds.top);
            if (!boundsArea)
                return Number.NaN;

            const viewportBounds = new DOMRect(window.visualViewport.offsetLeft, window.visualViewport.offsetTop, window.visualViewport.width, window.visualViewport.height);
            const maxLeft = Math.max(bounds.left, viewportBounds.left);
            const minRight = Math.min(bounds.right, viewportBounds.right);
            const maxTop = Math.max(bounds.top, viewportBounds.top);
            const minBottom = Math.min(bounds.bottom, viewportBounds.bottom);
            const intersectionArea = (minRight - maxLeft) * (minBottom - maxTop);
            return intersectionArea > 0 ? (100 * intersectionArea / boundsArea) : 0;
        }

        function checkFullscreenElement(event)
        {
            if (!fullscreenElement) 
                return;

            const visibility = calculateVisibilityPercentage(fullscreenElement.getBoundingClientRect());
            if (!visibility)
                logElementFailure(fullscreenElement, `'${event.type}' fired`);
        };

        function requestFullscreen(target)
        {
            fullscreenElement = target;
    
            observer = new IntersectionObserver(entries => {
                entries.forEach(entry => {
                    if (!entry.isIntersecting)
                        logElementFailure(entry.target, `IntersectionObserver called`);
                })
            });
            observer.observe(fullscreenElement);

            const fullscreenHandler = (event) => {
                if (document.fullscreenElement == fullscreenElement) {
                    setTimeout(() => { document.exitFullscreen() }, 100);
                    return;
                }
            
                fullscreenElement = null;
                setTimeout(() => {
                    event.target.removeEventListener('fullscreenchange', fullscreenHandler);
                    observer.disconnect();
                    observer = null;

                    testNextElement();
                });
            };

            waitForEvent('fullscreenchange', fullscreenHandler, false, false, target);

            target.requestFullscreen();
        }
            
        async function testElementById(id)
        {
            consoleWrite(`* Testing element '${id}'`);
            setTimeout(() => { 
                const element = document.getElementById(id);
                window.scrollTo(0, element.offsetTop);
    
                if (!internals)
                    return;
    
                setTimeout(() => {
                    internals.withUserGesture(() => { requestFullscreen(element.parentElement.parentElement) });
                }, 100);            
            }, 100);
        }
        
        const elementsToTest = ['first', 'second', 'third', 'fourth'];
        function testNextElement()
        {
            consoleWrite(``);

            const nextElementID = elementsToTest.splice(0, 1)[0];
            if (!nextElementID) {
                endTest();
                return;
            }

            setTimeout(() => { testElementById(nextElementID) }, 100);
        }
    
        window.addEventListener('load', () => {
            consoleElement = document.querySelector('#console');

            window.addEventListener('scroll', (event) => {
                if (fullscreenElement)
                    checkFullscreenElement(event);
            });

            testNextElement();
            
        });
    
    </script>
    
    </head>
    <body>
        <div id='console'></div>

        <div class='video-grandparent'>
            <div class='video-parent'>
                <div id='first' class='video-container'></div>
            </div>
        </div>
        <br>
        <div class='video-grandparent'>
            <div class='video-parent'>
                <div id='second' class='video-container'></div>
            </div>
        </div>
        <br>
        <div class='video-grandparent'>
            <div class='video-parent'>
                <div class='video-container'></div>
            </div>
        </div>
        <br>
        <div class='video-grandparent'>
            <div class='video-parent'>
                <div class='video-container'></div>
            </div>
        </div>
        <br>
        <div class='video-grandparent'>
            <div class='video-parent'>
                <div class='video-container'></div>
            </div>
        </div>
        <br>
        <div class='video-grandparent'>
            <div class='video-parent'>
                <div id='third' class='video-container'></div>
            </div>
        </div>
        <br>
        <div class='video-grandparent'>
            <div class='video-parent'>
                <div class='video-container'></div>
            </div>
        </div>
        <br>
        <div class='video-grandparent'>
            <div class='video-parent'>
                <div class='video-container'></div>
            </div>
        </div>
        <br>
        <div class='video-grandparent'>
            <div class='video-parent'>
                <div class='video-container'></div>
            </div>
        </div>
        <br>
        <div class='video-grandparent'>
            <div class='video-parent'>
                <div class='video-container'></div>
            </div>
        </div>
        <br>
        <div class='video-grandparent'>
            <div class='video-parent'>
                <div id='fourth' class='video-container'></div>
            </div>
        </div>
    </body>
</html>
